refactor coil API and split inductance into mutual, self, and filaments
